- Moved test cases around to correspond to TCM system - Updated some values to align with documented test case
